+++ server/src/services/stats.service.ts
@@ -1,3 +1,4 @@
+import { NextFunction } from 'express';
 import { Response } from 'express-serve-static-core';
 import { QueryResult } from 'pg';
 import StatsSearchResult from '../../../types/statistics.interfaces';
@@ -18,9 +19,10 @@ export const handleAddPomodoro = async (res: Response<Error | void, number>, log
 };
 
 const selectDate = "SELECT TO_CHAR(date, 'DD-MM-YYYY') as date, COUNT (date) FROM pomodoros WHERE userID = ($1) ";
+const groupAndOrder = 'GROUP BY date ORDER BY date DESC';
 
 export const handleGetAllStats = async (login: string, res: Response<StatsSearchResult>) => {
-  const sql = `${selectDate} GROUP BY date ORDER BY date DESC`;
+  const sql = `${selectDate} ${groupAndOrder}`;
 
   try {
     const queryResult: QueryResult = await client.query(sql, [login]);
@@ -31,30 +33,39 @@ export const handleGetAllStats = async (login: string, res: Response<StatsSearch
   }
 };
 
-export const getStatsFrom = async (req: Request, res: Response<StatsSearchResult>) => {
+export const getStatsInGivenMonth = async (req: Request, res: Response<StatsSearchResult>, next: NextFunction) => {
   const login = req.user.login;
-  const { from } = req.params;
-  const sql = `${selectDate} AND DATE >= ($2) GROUP BY date ORDER BY date DESC`;
+  const { date } = req.params;
+  const [year, month] = req.params.date.split('-').map((item: string) => Number(item));
+
+  if (!isYearCorrect(year)) {
+    console.log('error');
+
+    setError('year', year, next);
+  } else if (!isMonthCorrect(month)) {
+    setError('month', month, next);
+  } else {
+    await searchResultsInDb(login, date, res);
+  }
+};
+
+const searchResultsInDb = async (login: string, date: string, res: Response<StatsSearchResult, number>) => {
+  const sql = `${selectDate} and to_char(date, 'YYYY-MM') = ($2) ${groupAndOrder}`;
 
   try {
-    const queryResult = await client.query(sql, [login, from]);
+    const queryResult = await client.query(sql, [login, date]);
     res.json({ result: queryResult.rows });
   } catch (err) {
-    console.log(`handleGetPomodorosInPeriod${err.stack}`);
+    console.log(`err fetching getStatsInGivenMonth ${err}`);
     res.sendStatus(500).json(err.stack);
   }
 };
 
-export const getStatsInGivenPeriod = async (req: Request, res: Response<StatsSearchResult>) => {
-  const login = req.user.login;
-  const { from, to } = req.params;
-  const sql = `${selectDate} AND DATE BETWEEN ($2) AND ($3) GROUP BY date ORDER BY date DESC`;
-
-  try {
-    const queryResult = await client.query(sql, [login, from, to]);
-    res.json({ result: queryResult.rows });
-  } catch (err) {
-    console.log(`handleGetPomodorosInPeriod${err.stack}`);
-    res.sendStatus(500).json(err.stack);
-  }
+const setError = (value: 'year' | 'month', month: number, next: NextFunction) => {
+  const info = `Wrong ${value} ${month}`;
+  console.log(info);
+  next(info); // todo - add error handler
 };
+
+const isYearCorrect = (year: number): boolean => year >= 2020;
+const isMonthCorrect = (month: number): boolean => month > 0 && month < 13;