+++ changelog.md
@@ -1,3 +1,19 @@
-0.0.3 - handle stats in table
-0.0.2 - handle login error, rwd changes
-0.0.1 - adding login with password
+## 0.5.0 - 2021-01-30
+
+- api for stats for fetching months and years
+
+## 0.4.0 - 2021-01-29
+
+- moving to eslint
+
+## 0.3.0
+
+- handle stats in table
+
+## 0.2.0
+
+- handle login error, rwd changes
+
+## 0.1.0
+
+- adding login with password